Fixed SMD pad positions
Footprints on protoype PCBs dont match very well with real connectors.

Corrected SMD pad position according to TE, 3M, Amphenol and Harwin
documents, it should look like this:
http://www.mouser.com/ds/2/181/M50-310-350829.pdf
The only difference is that they recommend different pad length. The
current length of 2.1mm is somewhere in the middle of the
recommendations and will fit well for SMD processes. For hand soldering
they will be too short.